refactor(webapp): extract shared activity form utilities
Extract 7 duplicated helper functions into a single shared module:
- getDefaultDatetime, toSeconds, toMinutesSeconds
- formatDuration, formatTime, formatDate, toDateKey

Updated 5 page files to import from @/lib/activity-form-utils
instead of defining local copies. No logic changes.
